mipy-telegram/requirements.txt
Flask==2.0.1
librouteros==3.2.0
python-telegram-bot==13.7
python-dotenv==0.19.0
requests==2.26.0 